// backend/src/infrastructure/prisma/generated/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  ADMIN: 'ADMIN',
  DEVELOPER: 'DEVELOPER',
  SUPPORT: 'SUPPORT',
  MANAGER: 'MANAGER',
  QA: 'QA'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const TicketStatus = {
  OPEN: 'OPEN',
  IN_PROGRESS: 'IN_PROGRESS',
  RESOLVED: 'RESOLVED',
  CLOSED: 'CLOSED',
  CANCELLED: 'CANCELLED'
} as const

export type TicketStatus = (typeof TicketStatus)[keyof typeof TicketStatus]


export const TicketPriority = {
  LOW: 'LOW',
  MEDIUM: 'MEDIUM',
  HIGH: 'HIGH',
  URGENT: 'URGENT'
} as const

export type TicketPriority = (typeof TicketPriority)[keyof typeof TicketPriority]